Roll out GPU/CPU dispatch for all of dask-image
